Update schedule_tracker
Second fully working snippet of code, it fuctions properly, but it is not nice. I want to add to it a small GUI or something of that style, maybe host it on a website and add a java/html frontend. Not sure what these things mean yet, I will learn as I go.